cond_tree: an abstraction for expanding giant if trees to invert loops
In Synthizer's C++ version, we had an abstraction called vbool, which would allow us to write
normal loops and then wrap them in some boilerplate that would lift conditions out to constants and convert the entire
thing to a tree that expanded over all combimnations. THis in effect
allowed for writing normal loops which would then invert as if we had copy/pasted them
several times, moving all contained if statements up to instead surround the loop body. This is useful
for SIMD, as it can "prove" that a branch always has the same value when the compiler cnanot know because it is
based on runtime values.