dsp: add an OutputBuffer trait and some common implementations
This is useful to abstract over whether or not we are adding to an output, which is used to let everything avoid
mixing via intermediate buffers, since it can just write to the output and not have to care. Unfortunately
limitations of Rust mean that the code here is pretty repetitive.