synthizer/EdgeMap: rewrite to be O(log n) on read operations
This is `O(n log n)` to sort, then `O(log n)` after that. We use
small keys and sort_unstable, which means that we have for example 24*8192=196K.  This
fits comfortably in L2 for most practical cases.

We can revisit optimizing later: there are yet smarter structures. But the immediate goal is to get rid
of O(n^2) in the actual node graph, which is too impractical even for small examples.