Drop views and everything that depends on them
In C++, it makes sense to use interleaved channels.  This is fine in a language which
doesn't have a concept of safety, since there is no overhead to e.g. accessing slices
and you can do whatever.  in Rust, however, thereis much
gain to be had by storing samples un-interleaved, since this allows us
to work with [f32; BLOCK_SIZE] and similar.