Revert "Introduce support for enumerating devices"; cpal isn't as mature as miniaudio
This reverts commit 00acb40da5fb3f8964e57363bc633186514524fc.

We must use Miniaudio.  Cpal is immature and has bugs that I know about from reading their code:

- Wasapi:
  - Doesn't set thread priority.
  - Doesn't test for sample formats
  - Doesn't automatically route between devices if things are unplugged.
  - Doesn't expose stable device ids.
  - Uses crossbeam to communicate with the audio thread (probably not a big deal, but arguably unsafe).
  - Does things with COM initialization that are overly complex.
- Alsa/other Linux:
  - Doesn't handle surround sound channel formats, which are different on Linux.
- In general:
  - Contains a lot of todo.
  - Docs aren't very clear on what is and isn't supported.

Rust in general likes Cpal, and it's probably fine, but we can do better. If someone wants Webasm, we can give them webasm by using Cpal for that:
webasm is easy enough that it's kinda hard to get it wrong and they have a reasonable implementation of that.