Drop all our graph infrastructure
The problem with the old way of doing this was lack of flexibility w.r.t. cycle detection and
the like. We can do better with e.g. im, or even just
fully cloning (graphs are small), then optimize as needed (e.g. by "locking" the graph on
the writer threads, or whatever).