We now have a proper graph with inputs internally.
The minimal external API is the same, but nodes may have 0 or 1 inputs
or outputs, and we connect sine waves to an invisibly created output node that we maintain internally.

We must now turn our focus to figuring out handles.  Hopefully, at that point, the giant
death commits stop in favor of smaller, more targeted things.