Refactor to make the internal precision of Synthizer f64
This is slower, but necessary for automation/connecting to properties.  We can
optimize the slow things by converting to f32 locally or, in the future, using supermatch and specializing
on the cases in which we are aware that we only need f32 precision; it is much easier to go
from f64 to f32 peacemeal than it will be to chase down bugs if we try to go the other way (to see why, consider that
f32 to sometimes-f64 would hit the whole library, even fast-enough things, not just special cases like hrtf).