Refactor so that servers don't have multiple kinds
If we instead have the internal infrastructure give back a server command channel and a
callback which will run it, we can just call the callback instead.  This gets rid of the enum, and closes a number
of questions around whether or not users need to deal with different types due to send/sync bounds; the downside
is that we do now expose the command channel more directly (and, thus, it will need to learn to grow--but that can come later, 100k before blocking
is quite a lot).