Stop ignoring Cargo.lock
Rust is changing their advice on this in the near future, but even if they weren't, we value
reliable CI builds more than "look! It's always the latest deps!".  If it
gets to the point where it matters, we'll have GitHub actions run a cargo update in some jobs.

(From Cargo's perspective this changes nothing when on crates.io. The effect is merely that CI/other contributors no longer
have different versions of deps).