Bootstrap sample sources as far as resampling
They're now buffered, and they run through Rubato which does a good enough job to get us started. There is a fast
path wherein we don't resample if it's not needed.

Unfortunately testing is lacking.  We'll certainly circle back on that in the near future when
we start writing our testing framework, but as with other parts of this, we consider
it more important to get off the ground for the time being.  The long-term plan is golden master testing,
which will find such issues indirectly; testing this pipeline directly is arguably not worth it, but I haven't made a final decision on that yet. If nothing else
that would turn into testing Rubato, and it is unclear what Rubato's outputs should even be as the docs are exposing what look like they should be internal knobs and aren't very specific
on the algorithm.