integration_tests: remove the concept of multi-pass validators
We only have one special case we want to do with that so we will simply do it in the parent runners as a one-off later

This also gets rid of the golden validator because this history is already a mess so why not?  We'll be collapsing
this branch when we merge to master.