Move to using RPITIT instead of destination senders for signal inputs and outputs via a new `ValueProvider` trait
This should later let us borrow from traits and also allows signals to be lazy, which puts us in the position
we need to be in for the world of automation and heavier use of slots.  It also often
gets rid of stack space usage in debug builds; for example, scalars no longer expand to arrays and rely on LLVM
to notice that's what happened.

It's also just easier to read overall.

This flips from continuation passing style to call it "future"-style: signals can do work
early or late, where late may not happen.  Like how a future can be cancelled by being dropped.  Instead of
sandwiching the body of `tick` between the parent and child signal, `tick` builds the sandwich: early work, anything
child signals are going to do, possible late work if the providers get used by child signals.